aulas 02 pt2 - 03 - 04 - 05

+++ tdd/src/main/java/net/lorens/code/tdd/leilao/servico/MatematicaMaluca.java
@@ -2,7 +2,7 @@ package net.lorens.code.tdd.leilao.servico;
 
 public class MatematicaMaluca {
 
-    public int contaMaluca(int numero) {
+    public long contaMaluca(long numero) {
         if (numero > 30)
             return numero * 4;
         else if (numero > 10)
